refactor: improve settings handling
