+++ packages/shared/src/get-react-version.ts
@@ -5,7 +5,7 @@ import { match, P } from "ts-pattern";
 
 const _require = module.createRequire(import.meta.url);
 
-export function getReactVersion(fallback = "19.0.0"): string {
+export function getReactVersion(fallback: string): string {
   try {
     return match(_require("react"))
       .with({ version: P.select(P.string) }, identity)
