+++ packages/shared/src/index.ts
@@ -2,5 +2,4 @@ export * from "./constants";
 export * from "./get-doc-url";
 export * from "./get-id";
 export * from "./get-react-version";
-export * from "./schemas";
 export * from "./settings";
